requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.66.3
pandas==2.2.1
gunicorn==21.2.0
watchtower==3.0.1
boto3==1.34.69
diskcache==5.6.3
aiohttp==3.9.3
ratelimit==2.2.1 